requirements.txt
git+https://github.com/huggingface/transformers.git@v4.57.3
opencv_python_headless==4.9.0.80
opencv_python==4.8.0.74
huggingface_hub
open-clip-torch
beautifulsoup4
albumentations
sentencepiece 
numpy==1.26.4
torchmetrics
torchvision
mdtex2html
html2text
spaces
einops
gradio
pillow
torch
timm